+++ addons/web/static/src/js/core/abstract_service.js
@@ -2,6 +2,7 @@ odoo.define('web.AbstractService', function (require) {
 "use strict";
 
 var Class = require('web.Class');
+const { serviceRegistry } = require("web.core");
 var Mixins = require('web.mixins');
 var ServicesMixin = require('web.ServicesMixin');
 
@@ -40,6 +41,50 @@ var AbstractService = Class.extend(Mixins.EventDispatcherMixin, ServicesMixin, {
             this.env.bus.trigger('do-action', payload);
         }
     },
+
+    //--------------------------------------------------------------------------
+    // Static
+    //--------------------------------------------------------------------------
+
+    /**
+     * Deploy services in the env (specializations of AbstractService registered
+     * into the serviceRegistry).
+     *
+     * @static
+     * @param {Object} env
+     */
+    deployServices(env) {
+        const UndeployedServices = Object.assign({}, serviceRegistry.map);
+        function _deployServices() {
+            let done = false;
+            while (!done) {
+                // find a service with no missing dependency
+                const serviceName = Object.keys(UndeployedServices).find(serviceName => {
+                    const Service = UndeployedServices[serviceName];
+                    return Service.prototype.dependencies.every(depName => {
+                        return env.services[depName];
+                    });
+                });
+                if (serviceName) {
+                    const Service = UndeployedServices[serviceName];
+                    const service = new Service(env);
+                    env.services[serviceName] = service;
+                    delete UndeployedServices[serviceName];
+                    service.start();
+                } else {
+                    done = true;
+                }
+            }
+        }
+        serviceRegistry.onAdd((serviceName, Service) => {
+            if (serviceName in env.services || serviceName in UndeployedServices) {
+                throw new Error(`Service ${serviceName} is already loaded.`);
+            }
+            UndeployedServices[serviceName] = Service;
+            _deployServices();
+        });
+        _deployServices();
+    }
 });
 
 return AbstractService;
